publish impl for edges

+++ src/main/java/tektonikal/customblockhighlight/Vertexer.java
@@ -143,7 +143,7 @@ public class Vertexer {
         return in1;
     }
 
-    public static void vertexLine(MatrixStack matrices, BufferBuilder builder, float x1, float y1, float z1, float x2, float y2, float z2, Color cols, Color col2, int alpha, int nx, int ny, int nz) {
+    public static void vertexLine(MatrixStack matrices, BufferBuilder builder, float x1, float y1, float z1, float x2, float y2, float z2, Color cols, Color col2, int alpha, float nx, float ny, float nz) {
         Matrix4f model = matrices.peek().getPositionMatrix();
 
         builder.vertex(model, x1, y1, z1).color(cols.getRed(), cols.getGreen(), cols.getBlue(), alpha).normal(matrices.peek(), nx, ny, nz);
